refactor(adapter-openclaw): extract session activity + store cache into session-activity.ts
Seventh cut of the runtime.ts god-file split. Moves the session-transcript
activity layer into a new session-activity.ts: watchOpenClawSessionActivity /
createOpenClawSessionActivityCursor / readOpenClawSessionActivity (tail+cursor
read → ChatChunk stream), the mtime-guarded session-store cache (sessionStoreCache
+ SESSION_STORE_CACHE_MAX + readSessionStoreCached), the agent/sessionKey →
session-file resolution (resolveOpenClawSessionFile, findOpenClawSessionStoreEntry,
openClawCliSessionId, isOpenClawCliSessionId, deterministicUuid, readFileTail),
the 2 session types, and OPENCLAW_SESSION_ACTIVITY_POLL_MS. It imports the
transcript→ChatChunk builder from activity-summary (sibling).

The sessionStoreCache LRU cell now lives in exactly one module. The 3
__*ForTest cache accessors + SESSION_STORE_CACHE_MAX are RE-EXPORTED from
runtime.ts so session-store-cache.test.ts's `from '.../runtime'` import path
stays stable. The class imports back the 3 functions + poll const its stream
path uses; the rest stay module-internal.

Pure relocation — bodies byte-for-byte unchanged. runtime.ts shed its now-unused
createHash/readFileFrom/sleep/activityChunks imports. runtime.ts 2,156 -> 1,992.

Verified: typecheck, eslint (changed files), adapter suite 115-0 incl.
session-store-cache.test.ts + merge-chat-streams.test.ts (--isolate; the
re-exported test seams resolve), full suite 5124-0, binary build (stamps
reverted), boot smoke (adapter loads, schedule activates, list route 200).
